move permission class to separate file
